+++ tldextract/__init__.py
@@ -1,7 +1,7 @@
 """Export tldextract's public interface."""
 
 from . import _version
-from .tldextract import ExtractResult, TLDExtract, extract, update
+from .tldextract import ExtractResult, TLDExtract, extract, reverse_domain_name, update
 
 __version__: str = _version.version
 
@@ -10,5 +10,6 @@ __all__ = [
     "extract",
     "ExtractResult",
     "TLDExtract",
+    "reverse_domain_name",
     "update",
 ]