+++ tldextract/tldextract.py
@@ -41,6 +41,7 @@ import urllib.parse
 from collections.abc import Collection, Sequence
 from dataclasses import dataclass
 from functools import wraps
+from typing import Optional
 
 import idna
 import requests
@@ -518,3 +519,30 @@ def _decode_punycode(label: str) -> str:
         except (UnicodeError, IndexError):
             pass
     return lowered
+
+
+def reverse_domain_name(
+    domain: str,
+    tldextractor: Optional[TLDExtract] = None,
+) -> str:
+    """
+    An implementation of `Reverse Domain Name Notation`, in which the
+    registered domain is used as the leftmost component.
+
+    Reverse Domain Name Notation is typically used to organize namespaces for
+    packages and plugins.
+
+    >>> reverse_domain_name("login.example.com")
+    com.example.login
+
+    >>> reverse_domain_name("login.example.co.uk")
+    co.uk.example.login
+    """
+    if tldextractor:
+        result = tldextractor(domain)
+    else:
+        result = extract(domain)
+    stack = [result.suffix, result.domain]
+    if result.subdomain:
+        stack.extend(reversed(result.subdomain.split(".")))
+    return ".".join(stack)